+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## 0.8.2 (2020-03-12)
+
+BUG FIXES:
+
+- Fix `DescribeSubnetResource`
+
+## 0.8.1 (2020-03-11)
+
+ENHANCEMENTS:
+
+- Update `DescribeSubnet` api
+
+## 0.8.0 (2019-12-03)
+
+FEATURES:
+
+- Add `ufs` product (#42)
+
+## 0.7.0 (2019-11-21)
+
+FEATURES:
+
+- Add ssl options of client (#37)
+
+ENHANCEMENTS:
+
+- Add testing report use ucloud test framework (#38)
+
 ## 0.6.2 (2019-10-25)
 
 ENHANCEMENTS: